+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+2.5.0
+=====
+
+* (improvement) Only support React 19+.
+* (feature) Add `fetchApi()` helper.
+* (improvement) Remove dependency that can be built natively.
+
+
 2.4.0
 =====